Add suspense in categories lauout for CategoryName
