+++ components/GoodSmallCard.tsx
@@ -1,5 +1,3 @@
-import { cn } from '@/lib/utils';
-
 import {
   Card,
   CardContent,
